ReviewItem: A component that displays a single review for a product, including the reviewer's avatar, rating, name, and comment.
RelatedProducts: A component that displays related products in a slider format. It fetches related product data using the useGetRelatedProductsQuery hook and renders each related product as a ProductItem component.

ProductQuantity: A component that provides a quantity input for selecting the number of products to add to the cart. It uses Redux to manage the cart state and dispatches actions for incrementing and decrementing the quantity.

ProductDetailsCountdown: A component that displays a countdown timer for a flash sale. If the current date is after the offerExpiryTime, it displays a static countdown showing 0 days, hours, minutes, and seconds.

ProductDetailsArea: A higher-level component that integrates other components to create the product details page layout. It displays the product images, information, tabs (Description, Additional Information, and Reviews), and related products.

DetailsWrapper: A component that displays the main details of the product, such as its title, category, price, rating, description, and actions (add to cart, buy now, etc.). It also handles interactions related to the quantity, wishlist, compare, and product status.

DetailsThumbWrapper: A component responsible for rendering the product images and handling the selection of the active image. It also supports playing a video associated with the product.

DetailsTabNav: A component that renders the navigation tabs (Description, Additional Information, and Reviews) for the product details page. It also includes a dynamic marker that highlights the active tab.

DetailsBottomInfo: A component that displays additional information about the product, such as SKU, category, and tags. It also includes social media sharing options and information about the payment options.